# docs/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact:
    email: vaish@gmail.com
    name: API Support
  description: API documentation for the DJJS Event Reporting Backend.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: DJJS Event Reporting API
  version: "1.0"
paths:
...
  /api/files/{media_id}:
    delete:
      description: Deletes a file from S3 and optionally the media record. Optionally
        validates event_id to ensure file belongs to specific event.
      parameters:
      - description: Media ID
        in: path
        name: media_id
        required: true
        type: integer
      - description: Event ID (optional, for validation)
        in: query
        name: event_id
        type: integer
      - description: 'Delete media record from database (default: true)'
        in: query
        name: delete_record
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete file from S3
      tags:
      - Files
  /api/files/{media_id}/download:
    get:
      description: Generates a presigned URL for downloading a file from S3
      parameters:
      - description: Media ID
        in: path
        name: media_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get download URL for file
      tags:
      - Files
  /api/files/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload image, video, audio, or PDF file to S3 and associate with
        event media
      parameters:
      - description: File to upload (image, video, audio, or PDF)
        in: formData
        name: file
        required: true
        type: file
      - description: Event ID
        in: formData
        name: event_id
        required: true
        type: integer
      - description: Media ID (if updating existing media)
        in: formData
        name: media_id
        type: integer
      - description: File category (Event Photos, Video Coverage, Testimonials, Press
          Release)
        in: formData
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Upload file to S3
      tags:
      - Files
  /api/files/upload-multiple:
    post:
      consumes:
      - multipart/form-data
      description: Upload multiple image, video, audio, or PDF files to S3 and associate
        with event media
      parameters:
      - description: Files to upload (multiple files allowed)
        in: formData
        name: files
        required: true
        type: file
      - description: Event ID
        in: formData
        name: event_id
        required: true
        type: integer
      - description: File category (Event Photos, Video Coverage, Testimonials, Press
          Release)
        in: formData
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Upload multiple files to S3
      tags:
      - Files
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"